+++ suap_end_device/.vscode/settings.json
@@ -36,6 +36,7 @@
         "encoder.h": "c",
         "iot_button.h": "c",
         "task.h": "c",
-        "queue.h": "c"
+        "queue.h": "c",
+        "display.h": "c"
     }
 }
