+++ suap_end_device/main/drivers/display.c
@@ -1,0 +1,16 @@
+#include "display.h"
+
+SemaphoreHandle_t lvgl_mux = NULL;
+
+bool example_lvgl_lock(int timeout_ms)
+{
+    // Convert timeout in milliseconds to FreeRTOS ticks
+    // If `timeout_ms` is set to -1, the program will block until the condition is met
+    const TickType_t timeout_ticks = (timeout_ms == -1) ? portMAX_DELAY : pdMS_TO_TICKS(timeout_ms);
+    return xSemaphoreTakeRecursive(lvgl_mux, timeout_ticks) == pdTRUE;
+}
+
+void example_lvgl_unlock(void)
+{
+    xSemaphoreGiveRecursive(lvgl_mux);
+}
